# backend/docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      auth_token:
        type: string
      email:
        type: string
      id:
        type: string
      refresh_token:
        type: string
    type: object
  auth.RefreshTokenRequest:
    properties:
      token:
        type: string
    type: object
  auth.RefreshTokenResponse:
    properties:
      auth_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  btech-wallet_user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  server.ErrorResponse:
    properties:
      error: {}
      message:
        type: string
    type: object
  server.MessageResponse:
    properties:
      message:
        type: string
    type: object
  server.PaginationResponse-transaction_Transaction:
    properties:
      current_page:
        type: integer
      data:
        items:
          $ref: '#/definitions/transaction.Transaction'
        type: array
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  transaction.BalanceResponse:
    properties:
      balance:
        type: number
    type: object
  transaction.TopUpRequest:
    properties:
      amount:
        type: number
    type: object
  transaction.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      recipient_email:
        type: string
      recipient_id:
        type: string
      sender_email:
        type: string
      sender_id:
        type: string
    type: object
  transaction.TransferRequest:
    properties:
      amount:
        type: number
      to_user_email:
        type: string
    type: object
host: localhost:3333
info:
  contact: {}
  description: REST API for authentication and wallet transactions.
  title: BTech Wallet API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Returns a short-lived JWT and a refresh token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Authenticate a user
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Refresh an access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Register a user
      tags:
      - auth
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/btech-wallet_user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the current user profile
      tags:
      - profile
  /wallet/balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.BalanceResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wallet balance
      tags:
      - wallet
  /wallet/topup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Top-up details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.TopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top up wallet funds
      tags:
      - wallet
  /wallet/transaction/{transactionId}:
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: transactionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.Transaction'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a wallet transaction
      tags:
      - wallet
  /wallet/transactions:
    get:
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.PaginationResponse-transaction_Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List wallet transactions
      tags:
      - wallet
  /wallet/transfer:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transfer details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Transfer funds
      tags:
      - wallet
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"